+++ Sources/MapboxCoreNavigation/Navigator.swift
@@ -1,7 +1,8 @@
 import Foundation
 import MapboxNavigationNative
 
-extension Navigator {
+extension MapboxNavigationNative.Navigator {
+    
     func status(at timestamp: Date) -> NavigationStatus {
         return try! getStatusForMonotonicTimestampNanoseconds(
             Int64(timestamp.nanosecondsSince1970)